调整类路径 1、inlong-manager\manager-common\src\main\java\org\apache\inlong\manager\common\pojo\dataproxy的类迁移到inlong-manager\manager-dao\src\main\java\org\apache\inlong\manager\dao\entity 2、inlong-manager\manager-service\src\main\java\org\apache\inlong\manager\service\repository的类迁移到inlong-manager\manager-common\src\main\java\org\apache\inlong\manager\common\pojo\dataproxy
