+++ inlong-manager/manager-common/src/main/java/org/apache/inlong/manager/common/pojo/dataproxy/CacheClusterObject.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.service.repository;
+package org.apache.inlong.manager.common.pojo.dataproxy;
 
 import java.util.HashMap;
 import java.util.Map;
