+++ inlong-manager/manager-dao/src/main/java/org/apache/inlong/manager/dao/entity/FlumeSinkExt.java
@@ -15,14 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.common.pojo.dataproxy;
+package org.apache.inlong.manager.dao.entity;
 
 /**
- * CacheClusterExt
+ * FlumeSinkExt
  */
-public class CacheClusterExt {
-//    `cluster_name`        varchar(128) NOT NULL COMMENT 'CacheCluster name, English, numbers and underscore',
-    private String clusterName;
+public class FlumeSinkExt {
+//    `parent_name`        varchar(128) NOT NULL COMMENT 'FlumeSink name, English, numbers and underscore',
+    private String parentName;
+//    `set_name`            varchar(128) NOT NULL COMMENT 'ClusterSet name, English, numbers and underscore',
+    private String setName;
 //    `key_name`            varchar(64)  NOT NULL COMMENT 'Configuration item name',
     private String keyName;
 //    `key_value`           varchar(256)          DEFAULT NULL COMMENT 'The value of the configuration item',
@@ -31,21 +33,39 @@ public class CacheClusterExt {
     private int isDeleted;
 
     /**
-     * get clusterName
+     * get parentName
      * 
-     * @return the clusterName
+     * @return the parentName
      */
-    public String getClusterName() {
-        return clusterName;
+    public String getParentName() {
+        return parentName;
     }
 
     /**
-     * set clusterName
+     * set parentName
      * 
-     * @param clusterName the clusterName to set
+     * @param parentName the parentName to set
      */
-    public void setClusterName(String clusterName) {
-        this.clusterName = clusterName;
+    public void setParentName(String parentName) {
+        this.parentName = parentName;
+    }
+
+    /**
+     * get setName
+     * 
+     * @return the setName
+     */
+    public String getSetName() {
+        return setName;
+    }
+
+    /**
+     * set setName
+     * 
+     * @param setName the setName to set
+     */
+    public void setSetName(String setName) {
+        this.setName = setName;
     }
 
     /**
