+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/repository/DataProxyConfigRepository.java
@@ -28,19 +28,30 @@ import java.util.Timer;
 import java.util.TimerTask;
 
 import org.apache.commons.codec.digest.DigestUtils;
-import org.apache.inlong.manager.common.pojo.dataproxy.CacheCluster;
-import org.apache.inlong.manager.common.pojo.dataproxy.CacheClusterExt;
-import org.apache.inlong.manager.common.pojo.dataproxy.CacheTopic;
-import org.apache.inlong.manager.common.pojo.dataproxy.ClusterSet;
-import org.apache.inlong.manager.common.pojo.dataproxy.FlumeChannel;
-import org.apache.inlong.manager.common.pojo.dataproxy.FlumeChannelExt;
-import org.apache.inlong.manager.common.pojo.dataproxy.FlumeSink;
-import org.apache.inlong.manager.common.pojo.dataproxy.FlumeSinkExt;
-import org.apache.inlong.manager.common.pojo.dataproxy.FlumeSource;
-import org.apache.inlong.manager.common.pojo.dataproxy.FlumeSourceExt;
-import org.apache.inlong.manager.common.pojo.dataproxy.InLongId;
-import org.apache.inlong.manager.common.pojo.dataproxy.ProxyCluster;
-import org.apache.inlong.manager.common.pojo.dataproxy.ProxyClusterToCacheCluster;
+import org.apache.inlong.manager.common.pojo.dataproxy.CacheClusterObject;
+import org.apache.inlong.manager.common.pojo.dataproxy.CacheClusterSetObject;
+import org.apache.inlong.manager.common.pojo.dataproxy.CacheTopicObject;
+import org.apache.inlong.manager.common.pojo.dataproxy.DataProxyCluster;
+import org.apache.inlong.manager.common.pojo.dataproxy.DataProxyClusterSet;
+import org.apache.inlong.manager.common.pojo.dataproxy.DataProxyConfigResponse;
+import org.apache.inlong.manager.common.pojo.dataproxy.InLongIdObject;
+import org.apache.inlong.manager.common.pojo.dataproxy.ProxyChannel;
+import org.apache.inlong.manager.common.pojo.dataproxy.ProxyClusterObject;
+import org.apache.inlong.manager.common.pojo.dataproxy.ProxySink;
+import org.apache.inlong.manager.common.pojo.dataproxy.ProxySource;
+import org.apache.inlong.manager.dao.entity.CacheCluster;
+import org.apache.inlong.manager.dao.entity.CacheClusterExt;
+import org.apache.inlong.manager.dao.entity.CacheTopic;
+import org.apache.inlong.manager.dao.entity.ClusterSet;
+import org.apache.inlong.manager.dao.entity.FlumeChannel;
+import org.apache.inlong.manager.dao.entity.FlumeChannelExt;
+import org.apache.inlong.manager.dao.entity.FlumeSink;
+import org.apache.inlong.manager.dao.entity.FlumeSinkExt;
+import org.apache.inlong.manager.dao.entity.FlumeSource;
+import org.apache.inlong.manager.dao.entity.FlumeSourceExt;
+import org.apache.inlong.manager.dao.entity.InLongId;
+import org.apache.inlong.manager.dao.entity.ProxyCluster;
+import org.apache.inlong.manager.dao.entity.ProxyClusterToCacheCluster;
 import org.apache.inlong.manager.dao.mapper.ClusterSetMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
